# Optimize DateTime.Equals