A tiny improvement for `DateTime.Equals`:
```csharp
bool Test1(DateTime d1, DateTime d2) => d1 == d2;

// cmp against a constant
bool Test2(DateTime d1, DateTime d2) => d1 == new DateTime(637686893321649705);
```
Codegen diff: https://www.diffchecker.com/v4BfSbey
Could be done in JIT but it needs forward-sub feature for that first.

Self-contained benchmark: https://gist.github.com/EgorBo/e5642d9c2d362bb53880d3c9c87e5b95
```

|         Method |         d1 |         d2 |     Mean |    Error |   StdDev |
|--------------- |----------- |----------- |---------:|---------:|---------:|
|     Equals_new | MyDt[1000] | MyDt[1000] | 599.9 ns |  0.46 ns |  0.38 ns |
|     Equals_old | MyDt[1000] | MyDt[1000] | 651.8 ns |  0.70 ns |  0.62 ns |

|     Equals_new | MyDt[1000] | MyDt[1000] | 806.2 ns | 15.59 ns | 16.68 ns |
|     Equals_old | MyDt[1000] | MyDt[1000] | 859.2 ns |  2.34 ns |  2.19 ns |

| Equals_new_cns | MyDt[1000] | MyDt[1000] | 436.3 ns |  0.90 ns |  0.75 ns |
| Equals_old_cns | MyDt[1000] | MyDt[1000] | 855.7 ns | 13.78 ns | 12.89 ns |

| Equals_new_cns | MyDt[1000] | MyDt[1000] | 436.0 ns |  2.50 ns |  2.22 ns |
| Equals_old_cns | MyDt[1000] | MyDt[1000] | 875.0 ns | 16.63 ns | 17.80 ns |
```