@EgorBo 2021-10-01
hm... actually the benchmark is a bit flaky due to loop alignment. In case of "optimized" comparison the loop is aligned to 16 bytes.
The slower (main) code is more verbose so the loop is aligned to 32bytes naturally.
cc @kunalspathak