@jkotas 2021-10-01
What does it do to perf on 32-bit platforms? (I think it is ok to take this even it makes 32-bit slower, but it would be useful to have it as a datapoint.)