@jkotas 2021-10-01
Just a guess - masking the top two bits instead of shifting may be significantly faster on 32-bit, and about the same on 64-bit.