@EgorBo 2021-10-01
> What does it do to perf on 32-bit platforms? (I think it is ok to take this even it makes 32-bit slower, but it would be useful to have it as a datapoint.)

It seems it's still faster on 32bit, here is the codegen diff: https://www.diffchecker.com/vexK8vcb
```
|         Method |       Job |              Toolchain | IterationCount | LaunchCount | WarmupCount |         d1 |         d2 |    eq |       Mean |    Error |   StdDev |
|--------------- |---------- |----------------------- |--------------- |------------ |------------ |----------- |----------- |------ |-----------:|---------:|---------:|

// equals always return false
|     Equals_new | InProcess | InProcessEmitToolchain |        Default |     Default |     Default | MyDt[1000] | MyDt[1000] | False | 1,242.1 ns |  2.13 ns |  1.67 ns |
|     Equals_old | InProcess | InProcessEmitToolchain |        Default |     Default |     Default | MyDt[1000] | MyDt[1000] | False | 1,426.8 ns | 11.18 ns | 10.46 ns |

| Equals_new_cns | InProcess | InProcessEmitToolchain |        Default |     Default |     Default | MyDt[1000] | MyDt[1000] | False |   750.8 ns |  7.30 ns |  6.83 ns |
| Equals_old_cns | InProcess | InProcessEmitToolchain |        Default |     Default |     Default | MyDt[1000] | MyDt[1000] | False |   861.4 ns |  1.02 ns |  0.85 ns |


// equals always return true
|     Equals_new | InProcess | InProcessEmitToolchain |        Default |     Default |     Default | MyDt[1000] | MyDt[1000] |  True | 1,409.0 ns |  3.83 ns |  3.40 ns |
|     Equals_old | InProcess | InProcessEmitToolchain |        Default |     Default |     Default | MyDt[1000] | MyDt[1000] |  True | 1,605.4 ns | 13.36 ns | 11.16 ns |

| Equals_new_cns | InProcess | InProcessEmitToolchain |        Default |     Default |     Default | MyDt[1000] | MyDt[1000] |  True |   768.4 ns |  1.91 ns |  1.79 ns |
| Equals_old_cns | InProcess | InProcessEmitToolchain |        Default |     Default |     Default | MyDt[1000] | MyDt[1000] |  True |   861.5 ns |  1.81 ns |  1.60 ns |
```

> Just a guess - masking the top two bits instead of shifting may be significantly faster on 32-bit, and about the same on 64-bit.

Looks like it hurts x64 because ` << 2` 2 is contained and for the mask we have to request an additional reg. Diff for `Equals_new_cns` on x64: https://www.diffchecker.com/1pJUMY6t (`<<2` is on the left)