@kunalspathak 2021-10-01
> Self-contained benchmark: https://gist.github.com/EgorBo/e5642d9c2d362bb53880d3c9c87e5b95

Can you add this to dotnet/performance ?